orientation-event/DeviceMotion/values-null.html

<!DOCTYPE html>
<html>

  <head>
	<title>values set to null:</title>
	<script src="http://w3c-test.org/resources/testharness.js"></script>
	<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
	<script src="http://w3c-test.org/resources/WebIDLParser.js"></script>
	<script src="http://w3c-test.org/resources/idlharness.js"></script>
  <link rel='author' title='Mosquito FP7'/>
  </head>

  <body>

  
  	<h1>values set to null:</h1>
	<div id="log">
	</div>
	
	<script>
	
	function init(){
	   if (window.DeviceMotionEvent) {
			window.addEventListener('devicemotion', DeviceMotionHandler, false);
			alert("Put your device on a horizontal surface with the screen upmost.");
		};
	}
	
	function DeviceMotionHandler(eventData){
			
		var gravity = eventData.accelerationIncludingGravity;
		var rot = eventData.rotationRate;
		// X : angle par rapport a x
		x = gravity.x;
		// Y : angle par rapport a y
		y = gravity.y;
		// Z : acceleromètre (de base : -9.81)
		z = gravity.z;
		
		tx.step(function() {assert_equals(rot, null)});
		tx.done();

	}

	init();
	
	var tx = async_test("RotationRate set to null");
	</script>
	
  </body>
  
</html>
